topic count by bytes / sizeof(int32_t) not sizeof(long).
this is why if you typecast bytes to struct you use stdint.h not assume
sizeof(long) == 4.

+++ helpdeco.c
@@ -3845,7 +3845,7 @@ void ContextLoad(FILE *HelpFile)
     {
 	Topic=my_malloc(entries);
 	my_fread(Topic,entries,HelpFile);
-	Topics=(int)(entries/sizeof(long));
+	Topics=(int)(entries/sizeof(int32_t));
     }
 }
